// Client/src/types/member.ts
export type Member = {
  id: string
  email: string
  dateOfBirth: string
  imaageUrl?: string
  displayName: string
  created: string
  lastActive: string
  gender: string
  description: string
  city: string
  country: string
}

export type Photo= {
  id: number
  url: string
  publicId?: string
  memberId: string
}
export type EditableMember={
  displayName:string;
  description?:string;
  city:string;
  country:string;
}
export class MemberParams{
  gender?:string;
  minAge=18;
  maxAge=100;
  pageSize=10;
  pageNumber=1;
  orderBy='lastActive';
}